Update builtin-helpers.md
Hello, 
I saw that the Handlebars documentation was being updated, and I thought it was a good opportunity for me to give back to a project that I've loved. I've brought over the built-in helpers documentation from the current site. I'm not familiar with Vuepress, but I've tried to follow the format of other sections. I have some technical writing experience and am pretty familiar with Handlebars, so I'll try to help out where I can in the future. I see a few opportunities for making descriptions a little clearer, so I'll be back if the help is wanted.